Improves UI type safety and refactors facet logic
Introduces stricter TypeScript settings and documents required UI development practices to enforce strong type safety. Refactors facet sidebar logic to improve robustness, handling undefined and inactive facets more gracefully. Moves utility imports for better structure and maintainability. Extracts reusable components and reduces in-file component definitions for clarity.

Enhances keyboard navigation and autocomplete logic to guard against undefined or out-of-bounds indices, preventing runtime errors. Updates XML parsing and upload flows for safer attribute handling and more reliable processing of file uploads.

Overall, the changes aim to tighten type safety, improve user experience, and support maintainable, structured UI development.
