+++ ui/src/components/layout/index.ts
@@ -1,0 +1,3 @@
+export { default as ErrorBoundary } from './ErrorBoundary';
+export { default as Header } from './Header';
+export { default as Layout } from './Layout';
