+++ ui/src/components/search/FacetSidebar.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { createPortal } from 'react-dom';
 import type { FacetIndex } from '../../types';
 import { Filter, X, ChevronRight, Search as SearchIcon, Info } from 'lucide-react';
+import FacetPopover from './FacetPopover';
 
 interface FacetSidebarProps {
   facets: FacetIndex;
@@ -27,8 +28,10 @@ const FacetSidebar: React.FC<FacetSidebarProps> = ({
 
   const facetKeys = Object.keys(facets).sort((a, b) => {
     // Pin active facets to top
-    const aActive = facets[a].selected.size > 0;
-    const bActive = facets[b].selected.size > 0;
+    const aFacet = facets[a];
+    const bFacet = facets[b];
+    const aActive = aFacet ? aFacet.selected.size > 0 : false;
+    const bActive = bFacet ? bFacet.selected.size > 0 : false;
     if (aActive && !bActive) return -1;
     if (!aActive && bActive) return 1;
     return a.localeCompare(b);
@@ -113,8 +116,9 @@ const FacetSidebar: React.FC<FacetSidebarProps> = ({
       <div className="flex-1 overflow-y-auto">
         {filteredKeys.map(key => {
           const facet = facets[key];
+          if (!facet) return null;
           const isActive = facet.selected.size > 0;
-          
+
           return (
             <div key={key} className="relative">
               <button
@@ -137,21 +141,23 @@ const FacetSidebar: React.FC<FacetSidebarProps> = ({
                 </div>
               </button>
 
-              {openFacet === key && popoverPos && createPortal(
-                <FacetPopover
-                  title={key}
-                  facet={facet}
-                  onToggleValue={(val) => onToggleValue(key, val)}
-                  onSetMode={(mode) => onSetMode(key, mode)}
-                  onClear={() => onClearFacet(key)}
-                  style={{ 
-                    position: 'fixed', 
-                    top: Math.min(popoverPos.top, window.innerHeight - 400), 
-                    left: popoverPos.left + 4,
-                    zIndex: 9999 
-                  }}
-                />,
-                document.body
+              {openFacet === key && popoverPos && (
+                createPortal(
+                  <FacetPopover
+                    title={key}
+                    facet={facet}
+                    onToggleValue={(val) => onToggleValue(key, val)}
+                    onSetMode={(mode) => onSetMode(key, mode)}
+                    onClear={() => onClearFacet(key)}
+                    style={{
+                      position: 'fixed',
+                      top: Math.min(popoverPos.top, window.innerHeight - 400),
+                      left: popoverPos.left + 4,
+                      zIndex: 9999
+                    }}
+                  />,
+                  document.body
+                )
               )}
             </div>
           );
@@ -165,97 +171,6 @@ const FacetSidebar: React.FC<FacetSidebarProps> = ({
         >
           Clear All Filters
         </button>
-      </div>
-    </div>
-  );
-};
-
-interface FacetPopoverProps {
-  title: string;
-  facet: any;
-  onToggleValue: (val: string) => void;
-  onSetMode: (mode: 'any' | 'one') => void;
-  onClear: () => void;
-  style?: React.CSSProperties;
-}
-
-const FacetPopover: React.FC<FacetPopoverProps> = ({
-  title,
-  facet,
-  onToggleValue,
-  onSetMode,
-  onClear,
-  style
-}) => {
-  const [search, setSearch] = useState('');
-  
-  const filteredValues = facet.values.filter((v: any) => 
-    v.value.toLowerCase().includes(search.toLowerCase())
-  );
-
-  return (
-    <div 
-      className="facet-popover w-64 bg-white text-ink shadow-2xl rounded-md border border-steel overflow-hidden flex flex-col max-h-[400px]"
-      style={style}
-    >
-      <div className="p-3 border-b border-steel bg-paper flex items-center justify-between">
-        <h3 className="text-sm font-black uppercase tracking-tight truncate">{title}</h3>
-        <button onClick={onClear} className="text-[10px] font-bold text-ceremony hover:underline">Clear</button>
-      </div>
-
-      <div className="p-3 border-b border-steel grid grid-cols-2 gap-1">
-        <button
-          onClick={() => onSetMode('any')}
-          className={`text-[10px] font-bold uppercase py-1.5 rounded transition-colors ${
-            facet.mode === 'any' ? 'bg-ink text-paper' : 'bg-paper text-slate-500 hover:text-ink'
-          }`}
-        >
-          Include Any
-        </button>
-        <button
-          onClick={() => onSetMode('one')}
-          className={`text-[10px] font-bold uppercase py-1.5 rounded transition-colors ${
-            facet.mode === 'one' ? 'bg-ink text-paper' : 'bg-paper text-slate-500 hover:text-ink'
-          }`}
-        >
-          Require One
-        </button>
-      </div>
-
-      <div className="p-2">
-        <div className="relative">
-          <SearchIcon className="absolute left-2 top-1/2 -translate-y-1/2 w-3 h-3 text-slate-400" />
-          <input
-            type="text"
-            placeholder="Search values..."
-            value={search}
-            onChange={(e) => setSearch(e.target.value)}
-            className="w-full bg-paper border border-steel rounded px-7 py-1.5 text-xs focus:outline-none focus:border-ceremony/50"
-          />
-        </div>
-      </div>
-
-      <div className="flex-1 overflow-y-auto p-1">
-        {filteredValues.map((v: any) => (
-          <label
-            key={v.value}
-            className={`flex items-center gap-2 px-3 py-2 rounded text-sm cursor-pointer hover:bg-paper transition-colors ${
-              facet.selected.has(v.value) ? 'bg-ceremony/5 text-ceremony font-bold' : 'text-slate-600'
-            }`}
-          >
-            <input
-              type={facet.mode === 'one' ? 'radio' : 'checkbox'}
-              checked={facet.selected.has(v.value)}
-              onChange={() => onToggleValue(v.value)}
-              className="accent-ceremony"
-            />
-            <span className="truncate flex-1">{v.value}</span>
-            <span className="text-[10px] font-mono text-slate-400">({v.count})</span>
-          </label>
-        ))}
-        {filteredValues.length === 0 && (
-          <div className="p-4 text-center text-xs text-slate-400">No matching values</div>
-        )}
       </div>
     </div>
   );
