+++ ui/src/components/upload/index.ts
@@ -1,0 +1,1 @@
+export { default as BinRow } from './BinRow';
