+++ ui/src/services/catalogApi.ts
@@ -1,4 +1,4 @@
-import api from './api';
+import api from './apiClient';
 import type { 
   CatalogEntry, 
   Context,
