+++ ui/src/services/index.ts
@@ -1,0 +1,2 @@
+export { default as apiClient } from './apiClient';
+export { catalogApi } from './catalogApi';
